+++ src/main/java/org/nrg/xnat/dicomweb/plugin/DicomWebPlugin.java
@@ -6,13 +6,13 @@ import org.springframework.context.annotation.PropertySource;
 
 @PropertySource("classpath:/config/dicomweb/dicomweb.properties")
 @XnatPlugin(
-    value = "dicomwebproxy",
-    name = "DICOMweb Proxy Plugin",
+    value = "dicomwebplugin",
+    name = "DICOMweb Plugin",
     description = "Exposes XNAT projects as DICOMweb endpoints for OHIF and VolView",
     logConfigurationFile = "META-INF/resources/dicomweb-logback.xml",
     entityPackages = {"org.nrg.xnat.dicomweb", "org.nrg.xnatx.dicomweb.core.entity"},
     openUrls = {"/xapi/dicomweb/test"}
 )
 @ComponentScan({"org.nrg.xnat.dicomweb"})
-public class DicomWebProxyPlugin {
+public class DicomWebPlugin {
 }