+++ src/main/java/org/nrg/xnat/dicomweb/service/impl/StowRsServiceImpl.java
@@ -1,5 +1,5 @@
 /*
- * XNAT DICOMweb Proxy Plugin
+ * XNAT DICOMweb Plugin
  * Copyright (c) 2025 XNATWorks.
  * All rights reserved.
  */